Implement duration and QName constructors
