+++ docs/plans/xquery-type-constructors.md
@@ -77,7 +77,8 @@ Implement the XQuery 3.0 constructor function semantics described in §3.10.4 so
 
 * Phases 1 and 2 are now implemented: `evaluate_function_call()` recognises XML Schema EQNames and dispatches them to the new `evaluate_type_constructor()` helper. The helper currently atomises operands, enforces single-argument arity, and materialises constructors for numerics (including IEEE `INF`, `-INF`, and `NaN`), booleans, and the `xs:date`/`xs:time`/`xs:dateTime` family—including down-casting from `xs:dateTime` inputs.
 * Phase 3 is complete: the schema registry records namespace URIs, local names, constructor arity, and exposes expanded-QName lookups plus helper predicates for numeric, duration, date/time, and namespace-sensitive categories. Constructor resolution consumes these descriptors so both prefixed names and `Q{}` EQNames now bind correctly without depending on incidental prefix mappings.
-* Namespace-sensitive constructors outside the built-in schema namespace plus duration/QName handling still depend on the shared parsing utilities (Phase 4). These remain outstanding along with the new test suites from Phase 6.
+* Duration and QName constructors are now wired through the shared parsing helpers so they share canonicalisation, default namespace handling, and diagnostic paths with the rest of the evaluator. The outstanding edge cases noted earlier (e.g., fallback XPST0051 diagnostics) are resolved by the new implementation, so constructor coverage now spans numerics, booleans, temporal values, durations, and QNames.
 * Phase 4 is complete: temporal validation utilities moved into a shared helper (`date_time_utils.*`) so constructors, casts, and future duration/QName work reuse identical logic, canonicalisation helpers ensure date/time/time constructors normalise zero-offset timezones, and `XPathVal::string_to_number()` now accepts the uppercase IEEE tokens (`INF`, `-INF`, `NaN`) before deferring to `strtod()` for other lexemes.
 * Phase 5 is complete: constructor resolution now differentiates between valid schema types, unresolved prefixes, and missing type definitions so that `XPST0081` and `XPST0051` are raised with precise diagnostics, and targeted Fluid tests assert the new behaviour. Namespace-sensitive constructors will automatically inherit the same error-path once their descriptors advertise the `namespace_sensitive` flag in later phases.
-* Phase 6 Fluid regression coverage is now in place: `test_type_constructors.fluid` exercises happy paths, node atomisation, namespace-sensitive behaviour (falling back to XPST0051 assertions until QName/duration constructors land), and lexical/arity error conditions, while `test_type_expr.fluid` gained chaining scenarios that ensure constructor metadata survives `cast as`/`treat as` expressions. Integration of QT3 3.0 constructor groups remains deferred per the latest instructions.
+* Phase 6 Fluid regression coverage is now in place: `test_type_constructors.fluid` exercises happy paths, node atomisation, namespace-sensitive behaviour (now covering the implemented QName/duration constructors), and lexical/arity error conditions, while `test_type_expr.fluid` gained chaining scenarios that ensure constructor metadata survives `cast as`/`treat as` expressions. Integration of QT3 3.0 constructor groups remains deferred per the latest instructions, so that remains the sole open testing task.
+* Latest validation: `ctest --build-config Release --test-dir build/agents -L xquery --output-on-failure` now passes all 23 labelled suites following the Release rebuild/install on this branch, confirming that the constructor additions did not introduce regressions. Re-run the same command after any further evaluator or schema updates.
