+++ src/xml/schema/schema_types.h
@@ -38,6 +38,7 @@ namespace xml::schema
       XSInt,
       XSShort,
       XSByte,
+      XSQName,
       UserDefined
    };
